Split out base `Updater` class from PoularityUpdater
